Reusing activity as fixture
